Add up, down, and update commands to cmd.sh
Each command has been added for a different reason:

* `up`: docker-compose has some wonky behavior regarding how it deals
with configuration. I would like to move towards having a single unified
configuration file (or set of files) where the behavior of the node and
its services can be modified. In order to do that, due to
docker-compose's wonkiness, we're going to need a wrapper command. Might
as well make it `up`.

* `down`: To be symmetric with `up`.

* `update`: We don't have it in place now, but eventually there are
going to be backwards incompatible changes which need to be made which
will require more than a `git pull` (e.g. moving a user's configuration
files from A to B). Having an `update` command lets us have the ability
to automate these changes in the future.

This commit also changes the docs accordingly.